Tags in the sidebar will change style when active
