+++ _includes/sidebar.html
@@ -35,4 +35,4 @@
         {%- include social.html -%}
     </div>
 </div>
-<script src="/assets/script/sidebar.js"></script>
+<script type='module' src="/assets/script/sidebar.js"></script>
