+++ _sass/minima/_layout.scss
@@ -174,6 +174,11 @@
       background-color: $border-grey;
       color: white;
     }
+
+    .post--tag__active {
+      background-color: white;
+      color: $text-decoration-color;
+    }
   }
 
   .social-links {
