+++ assets/script/sidebar.js
@@ -1,3 +1,5 @@
+import { readTags } from "./tag-util.js";
+
 document.addEventListener('DOMContentLoaded', function() {
     const sessionStorage = window.sessionStorage;
     let tags = document.querySelectorAll('.sidebar--tags > .post--tag');
@@ -6,6 +8,23 @@ document.addEventListener('DOMContentLoaded', function() {
         sessionStorage.setItem('activatedTags', content);
         window.dispatchEvent(new Event("tagClicked"));
     }
+
+    const tagActiveClassName = 'post--tag__active';
+
+    const displayTagActiveStyle = function() {
+        let tagNames = new Set(readTags(sessionStorage));
+        for (let tag of tags) {
+            if (tagNames.has(tag.textContent)) {
+                tag.classList.add(tagActiveClassName);
+            }
+            else {
+                tag.classList.remove(tagActiveClassName)
+            }
+        }
+    }
+
+    window.addEventListener('load', displayTagActiveStyle);
+    window.addEventListener('tagClicked', displayTagActiveStyle);
     
     for (let tag of tags) {
         tag.addEventListener('click', function() {
@@ -29,5 +48,4 @@ document.addEventListener('DOMContentLoaded', function() {
             setSessionStorage(activatedTags);
         });
     }
-    // TODO: deactivate tag when it's clicked the second time
 });
