+++ pages/pertag.html
@@ -16,4 +16,4 @@ permalink: /tag/
 {%- endfor -%}
     </ul>
 </div>
-<script src="/assets/script/tags.js"></script>
+<script type='module' src="/assets/script/tags.js"></script>
